+++ src/properties/logger.ts
@@ -50,6 +50,8 @@ export default {
   foundBackupVaults: (num: number): string => `Found ${num} Backup Vaults`,
   foundBackupInstances: (num: number): string => `Found ${num} Backup Instances`,
   foundBackupPolicies: (num: number): string => `Found ${num} Backup Policies`,
+  /* Cognitive services account */
+  foundCognitiveServicesAccounts: (num: number): string => `Found ${num} cognitive services accounts`,
   /* CDN Profiles */
   foundCdnProfiles: (num: number): string => `Found ${num} CDN profiles`,
   /* CDN Endpoints */
@@ -141,8 +143,12 @@ export default {
   foundRedisCaches: (num: number): string => `Found ${num} Redis caches`,
   // Replication Appliances
   foundReplicationAppliances: (num: number): string => `Found ${num} replication appliances`,
+  // Replication Centers
+  foundReplicationCenters: (num: number): string => `Found ${num} replication centers`,
   // Replication Networks
   foundReplicationNetworks: (num: number): string => `Found ${num} replication networks`,
+  // Replication Policies
+  foundReplicationPolicies: (num: number): string => `Found ${num} replication policies`,
   // Resource Groups
   foundResourceGroups: (num: number): string => `Found ${num} resource groups`,
   // Security Assesments