+++ src/services/tag/connections.ts
@@ -1047,7 +1047,7 @@ export default ({
         }
       }
     }
-  
+
     /**
      * Find related backup vaults
      */
@@ -1093,6 +1093,31 @@ export default ({
             resourceType: services.recoveryVaults,
             relation: 'child',
             field: 'recoveryVaults',
+          })
+        }
+      }
+    }
+
+    /**
+     * Find related cognitive services account
+     */
+    const cognitiveServicesAccountList: {
+      name: string
+      data: { [property: string]: any[] }
+    } = data.find(({ name }) => name === services.cognitiveServicesAccount)
+    if (cognitiveServicesAccountList?.data?.[region]) {
+      const dataAtRegion: any = findServiceInstancesWithTag(
+        tag,
+        cognitiveServicesAccountList.data[region]
+      )
+      if (!isEmpty(dataAtRegion)) {
+        for (const cognitiveServicesAccount of dataAtRegion) {
+          const { id } = cognitiveServicesAccount
+          connections.push({
+            id,
+            resourceType: services.cognitiveServicesAccount,
+            relation: 'child',
+            field: 'cognitiveServicesAccounts',
           })
         }
       }
@@ -1173,7 +1198,7 @@ export default ({
       }
     }
   }
-  
+
   const tagResult = {
     [tag.id]: connections,
   }